Batched neural net evaluations
Group evaluations and run them in parallel. Roughly 50% speedup on my setup, but there are a couple of points that is debatable.

- Thread / batch sizing heuristics : This PR changes how the default threads / default batch sizes are picked.  See Leela.cpp
- Batch-forming heuristic : See OpenCLScheduler.cpp for the batch forming heuristic : the heuristic exists so that we can wait for the rest of the engine to create more NN evaluations so that we can run larger batches.  We can't wait indefinitely since there are cases we enter 'serial' paths.  Since heuristics are heuristics, these might need some tests on a larger variety of types of systems.

Did make sure that winrate improves when running default vs. default command line `./leelaz -w (weight file)` on time parity.

Pull request #2188.

+++ src/SMP.cpp
@@ -73,6 +73,6 @@ SMP::Lock::~Lock() {
     }
 }
 
-int SMP::get_num_cpus() {
+size_t SMP::get_num_cpus() {
     return std::thread::hardware_concurrency();
 }
